+++ src/objectwizards/wizards/alTableExtWizard.ts
@@ -19,13 +19,15 @@ export class ALTableExtWizard extends ALObjectWizard {
     }
 
     protected async runAsync(settings: ALObjectWizardSettings) {
-        let alLangProxy : ALLangServerProxy = new ALLangServerProxy();
+        let alLangProxy : ALLangServerProxy = this._toolsExtensionContext.alLangProxy;
         let objectId : string = await alLangProxy.getNextObjectId(settings.getDestDirectoryUri(), "tableextension");
 
         let wizardData : ALTableExtWizardData = new ALTableExtWizardData();
         wizardData.objectId = objectId;
         wizardData.objectName = '';
         wizardData.selectedTable = '';
+        wizardData.idRangeStart = 
+            alLangProxy.getIdRangeStart(settings.getDestDirectoryUri());
         let wizardPage : ALTableExtWizardPage = new ALTableExtWizardPage(this._toolsExtensionContext, settings, wizardData);
         wizardPage.show();
     }